% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.build_tree.R
\name{ds.build_tree}
\alias{ds.build_tree}
\title{Title}
\usage{
ds.build_tree(object, objective_column, questions = NULL, datasources = NULL)
}
\arguments{
\item{object}{basis of build tree}

\item{objective_column}{objective columns needed to build tree}

\item{questions}{questions needed to build tree}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
the built tree
}
\description{
Title
}
