% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.center.R
\name{ds.center}
\alias{ds.center}
\title{Centering of a data frame on the study server}
\usage{
ds.center(x, type = "combined", name = NULL, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the data frame (or numeric vector, \code{table$column}) on the study
server to be centered}

\item{type}{\code{type} (default \code{"combined"}) the centering is done with a pooled approach, using the
mean of the pooled data, if \code{"split"} the data will be centered with the mean of each server data}

\item{name}{\code{"character"} (default \code{NULL}) If \code{NULL}, the original data frame will be overwritten,
otherwise the centered data frame will be assigned to a variable named after this argument}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates (or overwrites) a data frame on the study server.
}
\description{
Centers each column of a data frame object by substracting the mean of the column
}
