% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.dummies.R
\name{ds.dummies}
\alias{ds.dummies}
\title{Create dummy variables}
\usage{
ds.dummies(x, name = NULL, prob_adjust = FALSE, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the data frame on the study server}

\item{name}{\code{character} (default \code{NULL}) Overwrites the data frame if \code{NULL},
otherwise the new data frame with dummies is asigned to a new variable}

\item{prob_adjust}{\code{bool} (default \code{FALSE}) Scale the dummy variables dividing them
by the square root of their probability (n/n_total)}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates (or overwrites) a data frame on the study server.
}
\description{
Create dummy variables for factor columns on server data frame
}
