% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.dummy_probability.R
\name{ds.dummy_probability}
\alias{ds.dummy_probability}
\title{Scale dummy variables}
\usage{
ds.dummy_probability(x, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the data frame on the study server}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates (or overwrites) a data frame on the study server.
}
\description{
Internal function. Scale using the square root of their probabilitiy.
}
