% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.FAMD.R
\name{ds.FAMD}
\alias{ds.FAMD}
\title{Factor Analysis on Mixed Data}
\usage{
ds.FAMD(x, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the data frame on the study server}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
\code{data frame} With the principal components (columns) and variables (rows)
}
\description{
Perform FAMD analysis on server data
}
