% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.find_best_split.R
\name{ds.find_best_split}
\alias{ds.find_best_split}
\title{Title}
\usage{
ds.find_best_split(object, objective_column, datasources = NULL)
}
\arguments{
\item{object}{}

\item{objective_column}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
