% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.PCA.R
\name{ds.PCA}
\alias{ds.PCA}
\title{Principal component analysis (PCA)}
\usage{
ds.PCA(
  x,
  method = "fast",
  ncomp = 2,
  scale = TRUE,
  center = TRUE,
  datasources = NULL
)
}
\arguments{
\item{x}{\code{character} Name of the data frame on the server that has the dataset}

\item{method}{\code{character} (default \code{"fast"}) Method of the SVD calculation. \code{"fast"} to do a truncated SVD of the first \code{ncomp}
components, \code{"full"} to do the complete SVD calculation.}

\item{ncomp}{\code{numeric} (default \code{2}) Number of components to calculate when \code{method = "fast"}}

\item{scale}{\code{bool} (default \code{TRUE}) Set to TRUE to scale the dataset (pooled approach), FALSE to not scale}

\item{center}{\code{bool} (default \code{TRUE}) Set to TRUE to center the dataset (pooled approach), FALSE to not scale}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
\code{data frame} With the principal components (columns) and variables (rows)
}
\description{
Perform a non-disclosive PCA with the option of centering and scaling (unit variance) the pooled data
}
