% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.scale.R
\name{ds.scale}
\alias{ds.scale}
\title{Scaling of a server side data frame or numeric vector}
\usage{
ds.scale(x, type = "combined", name = NULL, datasources = NULL)
}
\arguments{
\item{x}{\code{data.frame} or \code{numeric} vector to be scaled}

\item{type}{\code{type} (default \code{"combined"}) the scaling is done with a pooled approach, using the
standard deviation of the pooled data, if \code{"split"} the data will be scaled with the standard deviation of each server data}

\item{name}{\code{"character"} (default \code{NULL}) If \code{NULL}, the original data frame will be overwritten,
otherwise the scaled data frame will be assigned to a variable named after this argument}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates (or overwrites) a data frame on the study server.
}
\description{
Scales by dividing each variable by the standard deviation to have unitary variance
}
