% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.scatterPlotGG.R
\name{ds.scatterPlotGG}
\alias{ds.scatterPlotGG}
\title{Generates non-disclosive scatter plots (with \code{ggplot2})}
\usage{
ds.scatterPlotGG(
  x = NULL,
  y = NULL,
  group = NULL,
  method = "deterministic",
  k = 3,
  noise = 0.25,
  type = "split",
  datasources = NULL
)
}
\arguments{
\item{x}{\code{character} (default \code{NULL}) specifying the name of a numeric vector.}

\item{y}{\code{character} (default \code{NULL}) specifying the name of a numeric vector.}

\item{group}{\code{character} (default \code{NULL}) specifying the name of a numeric vector.}

\item{method}{\code{character} (default \code{'deterministic'}) string that specifies the method that is used to
generated non-disclosive coordinates to be displayed in a scatter plot. This argument can be set as \code{'deteministic'}
or \code{'probabilistic'}. For more information see Details.}

\item{k}{\code{numeric} (default \code{3}) he number of the nearest neighbors for which their centroid is calculated.
For more information see Details.}

\item{noise}{\code{numeric} (default \code{0.25}) the percentage of the initial variance that is used as the variance
of the embedded noise if the argument method is set to \code{'probabilistic'}.}

\item{type}{\code{character} (default \code{'split'}) a character that represents the type of graph to display.
This can be set as \code{'combine'} or \code{'split'}. For more information see Details.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
\code{ggplot} object
}
\description{
This function uses two disclosure control methods to generate non-disclosive scatter
plots of two server-side continuous variables, with the option of using a grouping factor variable.
}
\details{
As the generation of a scatter plot from original data is disclosive and is not
permitted in DataSHIELD, this function allows the user to plot non-disclosive scatter plots.

If the argument \code{method} is set to \code{'deterministic'}, the server-side function searches
for the \code{k-1} nearest neighbors of each single data point and calculates the centroid
of such \code{k} points.
The proximity is defined by the minimum Euclidean distances of z-score transformed data.

When the coordinates of all centroids are estimated the function applies scaling to expand the
centroids back to the dispersion of the original data. The scaling is achieved by multiplying
the centroids with a scaling factor that is equal to the ratio between the standard deviation of
the original variable and the standard deviation of the calculated centroids. The coordinates of
the scaled centroids are then returned to the client-side.

The value of \code{k} is specified by the user.
The suggested and default value is equal to 3 which is also
the suggested minimum threshold that is used to prevent disclosure which is specified in the
protection filter \code{nfilter.kNN}. When the value of \code{k} increases,
the disclosure risk decreases but the utility loss increases.
The value of \code{k} is used only
if the argument \code{method} is set to \code{'deterministic'}.
Any value of \code{k} is ignored if the
argument \code{method} is set to \code{'probabilistic'}.

If the argument \code{method} is set to \code{'probabilistic'},
the server-side function generates a random normal noise of zero mean
and variance equal to 10\\% of the variance of each \code{x} and \code{y} variable.
The noise is added to each \code{x} and \code{y} variable and the disturbed by the addition of
\code{noise} data are returned to the client-side. Note that the seed random number generator is fixed to a
specific number generated from the data and therefore the user gets the same figure every time
that chooses the probabilistic method in a given set of variables.
The value of \code{noise} is used only if the argument \code{method} is set to \code{'probabilistic'}.
Any value of \code{noise} is ignored if
the argument \code{method} is set to \code{'deterministic'}.

In \code{type} argument can be set two graphics to display:\cr
(1) If \code{type = 'combine'}  a scatter plot for
combined data is generated.\cr
(2) If \code{type = 'split'}  one scatter plot for each
study is generated.

Server function called: \code{scatterPlotGGDS}
}
