% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.sd.R
\name{ds.sd}
\alias{ds.sd}
\title{Standard Deviation}
\usage{
ds.sd(x, type = "split", checks = FALSE, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the numeric vector on the server side to obtain it's SD}

\item{type}{\code{character} (default \code{"split"}) SD for each server (\code{"split"}) or
pooled approach (\code{"combined"})}

\item{checks}{\code{bool} (default \code{FALSE}) If \code{TRUE} optional checks of model components will be undertaken.
It is suggested that checks should only be undertaken once the function call has failed.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
\code{list} including:\cr

\code{Standard_deviation.by.Study}: estimated standard deviation, \code{Nmissing}
(number of missing observations), \code{Nvalid} (number of valid observations) and
\code{Ntotal} (sum of missing and valid observations)
separately for each study (if \code{type = split}).\cr
\code{Global.Standard_deviation}: estimated standard deviation, \code{Nmissing}, \code{Nvalid} and \code{Ntotal}
across all studies combined (if \code{type = combine}). \cr
\code{Nstudies}: number of studies being analysed. \cr
\code{ValidityMessage}: indicates if the analysis was possible. \cr
}
\description{
This function computes the standard deviation of the values of a server numerical vector.
The SD can be calculated server-wise or pooled
}
