% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.svd.R
\name{ds.svd}
\alias{ds.svd}
\title{Singular value decomposition (SVD)}
\usage{
ds.svd(x, method.block = "fast", ncomp = 2, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the data frame on the server that has the dataset}

\item{method.block}{\code{character} (default \code{"fast"}) Method of the SVD calculation. \code{"fast"}
to do a truncated SVD of the first \code{ncomp} components, \code{"full"} to do the complete SVD calculation}

\item{ncomp}{\code{numeric} (default \code{2}) Number of components to calculate when \code{method = "fast"}}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
\code{list} with: \cr
-d \code{numeric}: singular values of x, sorted decreasingly\cr
-u \code{numeric data frame}:left singular vectors\cr
}
\description{
Non-disclosive SVD on pooled data
}
