% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{leaf}
\alias{leaf}
\title{Title}
\usage{
leaf(object, questions, objective_column, datasources)
}
\arguments{
\item{object}{}

\item{questions}{}

\item{objective_column}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
