% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.build_tree.R
\name{ds.build_tree}
\alias{ds.build_tree}
\title{Title}
\usage{
ds.build_tree(object, objective_column, questions = NULL, datasources = NULL)
}
\arguments{
\item{object}{}

\item{objective_column}{}

\item{questions}{}

\item{datasources}{}
}
\description{
Title
}
