% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gini_index.R
\name{ds.gini_index}
\alias{ds.gini_index}
\title{Title}
\usage{
ds.gini_index(object, column, objective_column, datasources = NULL)
}
\arguments{
\item{object}{}

\item{column}{}

\item{objective_column}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
