% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_tree.R
\name{prepare_dataset_tree}
\alias{prepare_dataset_tree}
\title{Title}
\usage{
prepare_dataset_tree(object, objective_column, questions, datasources)
}
\arguments{
\item{object}{}

\item{questions}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
